refactor(core): share the query-group subscriber lookup
The query existed in two identical copies, one in the regeneration impact
resolver and one in the computed-attribute tasks. A third consumer is coming on
the merge path, and importing either copy from there would close an import loop.
